update-status-of-apis-in-openapi-examples/pyproject.toml
[tool.poetry]
name = "update-status-of-apis-in-openapi-examples"
version = "0.0.1"
description = ""
authors = ["dphuang2 <dylan@konfigthis.com>"]
readme = "README.md"
packages = [{include = "update_status_of_apis_in_openapi_examples"}]

[tool.poetry.dependencies]
python = "^3.9"
matplotlib = "^3.8.3"
pyyaml = "^6.0.1"
pandas = "^2.2.1"
colorlog = "^6.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
